requirements.txt
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
Pillow==2.7.0
PyYAML==3.10
Pygments==1.6
argparse==1.2.1
commando==0.3.4
cssprefixer==1.4.1
cssutils==1.0
fswrap==0.1.2
glue==0.9.4
hyde==0.8.8
python-dateutil==1.5
python-magic==0.4.3
smartypants==1.6.0.3
typogrify==2.0.0
wsgiref==0.1.2
